content polished

+++ trunk/Project/particleSystem.h
@@ -1,4 +1,4 @@
-/* particleSystem.h
+/* class ParticleSystem
 This class contains functions that allow a set of randomly moving particles to be generated. */
 
 #ifndef PARTICLESYSTEM_H
@@ -14,11 +14,12 @@ using namespace std;
 class ParticleSystem {
 
 private:
+	// Container for the Particle objects handled by this object
 	vector<Particle> particleList;
 
 public:
 	ParticleSystem();
-	void emit(D3DXVECTOR3 pos, int type_, int life_, float speed_, D3DXVECTOR3 direction);
+	void emit(D3DXVECTOR3 pos, int type_, float life_, float speed_, D3DXVECTOR3 direction);
 	void update();
 	vector<Particle> getParticles();
 };
